requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
langchain
langgraph
langchain-google-genai
faiss-cpu
langchain-community
rank_bm25
py2neo
python-dotenv
neo4j
pandas
python-multipart
pydantic
matplotlib
seaborn
fpdf
pathlib
twilio